refactor(eng-analytics): split presentation views by domain
